/set_profile now lets you upload an images to the channel instead of requiring a specifically formatted link
