+++ packages/game/src/commands/index.ts
@@ -21,7 +21,7 @@ import newgame from '@adventure-bot/game/commands/newgame'
 // import list from '@adventure-bot/game/commands/list/list'
 import quests from '@adventure-bot/game/commands/quests'
 // import renew from '@adventure-bot/game/commands/renew'
-import set from '@adventure-bot/game/commands/set'
+import setProfile from '@adventure-bot/game/commands/setProfile'
 import { CommandHandler } from '@adventure-bot/game/utils'
 
 const commands = new Map<
@@ -35,14 +35,11 @@ const commands = new Map<
 commands.set('adventure', adventure)
 commands.set('attack', attack)
 commands.set('heal', heal)
-// commands.set('d20', d20)
 commands.set('hp', hp)
 commands.set('inspect', inspect)
 commands.set('inventory', inventory)
-// commands.set('list', list)
 commands.set('quests', quests)
-// commands.set('renew', renew)
-commands.set('set', set)
+commands.set('set_profile', setProfile)
 commands.set('newgame', newgame)
 commands.set('admin', admin)
